ScaleInvariantSDR
