+++ README.md
@@ -266,7 +266,7 @@ acc = torchmetrics.functional.accuracy(preds, target)
 We currently have implemented metrics within the following domains:
 
 - Audio (
-  [SI_SDR](https://torchmetrics.readthedocs.io/en/latest/references/modules.html#si-sdr),
+  [ScaleInvariantSDR](https://torchmetrics.readthedocs.io/en/latest/references/modules.html#ScaleInvariantSDR),
   [SI_SNR](https://torchmetrics.readthedocs.io/en/latest/references/modules.html#si-snr),
   [SNR](https://torchmetrics.readthedocs.io/en/latest/references/modules.html#snr)
   and [few more](https://torchmetrics.readthedocs.io/en/latest/references/modules.html#audio-metrics)
