+++ tests/audio/test_si_sdr.py
@@ -21,7 +21,7 @@ from torch import Tensor
 
 from tests.helpers import seed_all
 from tests.helpers.testers import BATCH_SIZE, NUM_BATCHES, MetricTester
-from torchmetrics.audio import SI_SDR
+from torchmetrics.audio import ScaleInvariantSDR
 from torchmetrics.functional import si_sdr
 from torchmetrics.utilities.imports import _TORCH_GREATER_EQUAL_1_6
 
@@ -84,7 +84,7 @@ class TestSISDR(MetricTester):
             ddp,
             preds,
             target,
-            SI_SDR,
+            ScaleInvariantSDR,
             sk_metric=partial(average_metric, metric_func=sk_metric),
             dist_sync_on_step=dist_sync_on_step,
             metric_args=dict(zero_mean=zero_mean),
@@ -103,7 +103,7 @@ class TestSISDR(MetricTester):
         self.run_differentiability_test(
             preds=preds,
             target=target,
-            metric_module=SI_SDR,
+            metric_module=ScaleInvariantSDR,
             metric_functional=si_sdr,
             metric_args={"zero_mean": zero_mean},
         )
@@ -119,13 +119,13 @@ class TestSISDR(MetricTester):
         self.run_precision_test_gpu(
             preds=preds,
             target=target,
-            metric_module=SI_SDR,
+            metric_module=ScaleInvariantSDR,
             metric_functional=si_sdr,
             metric_args={"zero_mean": zero_mean},
         )
 
 
-def test_error_on_different_shape(metric_class=SI_SDR):
+def test_error_on_different_shape(metric_class=ScaleInvariantSDR):
     metric = metric_class()
     with pytest.raises(RuntimeError, match="Predictions and targets are expected to have the same shape"):
         metric(torch.randn(100), torch.randn(50))
